requirements.txt
fastapi
uvicorn
alembic 
sqlalchemy
sqlalchemy-utils
python-slugify
jinja2
